Fixed inventory and discardning of items
