+++ .gitignore
@@ -1,4 +1,3 @@
 player.json
-save_data.json
 /test
 __pycache__/
